# [DO NOT MERGE] ModelWriter implementation perf comparison