--- a/sdk/core/Azure.Core/src/Serialization/ModelWriterInterlocked.SequenceBuilder.cs
+++ b/sdk/core/Azure.Core/src/Serialization/ModelWriterInterlocked.SequenceBuilder.cs
@@ -158,7 +158,18 @@ namespace Azure.Core.Serialization
                 for (int i = 0; i < _count; i++)
                 {
                     Buffer buffer = _buffers[i];
-                    await stream.WriteAsync(buffer.Array, 0, buffer.Written).ConfigureAwait(false);
+                    await stream.WriteAsync(buffer.Array, 0, buffer.Written, cancellation).ConfigureAwait(false);
+                }
+            }
+
+            public void CopyToSpan(in Span<byte> bytes)
+            {
+                var span = bytes;
+                for (int i = 0; i < _count; i++)
+                {
+                    Buffer buffer = _buffers[i];
+                    buffer.Array.AsSpan(0, buffer.Written).CopyTo(span);
+                    span = span.Slice(buffer.Written);
                 }
             }
         }