… sdk/core/Azure.Core/tests/ModelSerialization/ModelWriterInterlockedTests.cs
… sdk/core/Azure.Core/tests/ModelSerialization/ModelWriterLocksTests.cs