```
AvailabilitySetDataModel:
|                            Method | SegmentSize |       Mean |     Error |    StdDev |     Median | Ratio |   Gen0 | Allocated | Alloc Ratio |
|---------------------------------- |------------ |-----------:|----------:|----------:|-----------:|------:|-------:|----------:|------------:|
|     Serialize_ModelWriter_Current |        4096 | 3,255.3 ns | 103.05 ns | 118.67 ns | 3,173.0 ns |  1.00 |      - |     728 B |        1.00 |
| Serialize_ModelWriter_Interlocked |        4096 |   741.0 ns |  17.65 ns |  20.32 ns |   728.8 ns |  0.23 | 0.0062 |     592 B |        0.81 |
|       Serialize_ModelWriter_Locks |        4096 |   786.4 ns |  15.59 ns |  17.33 ns |   778.2 ns |  0.24 | 0.0064 |     624 B |        0.86 |
|                                   |             |            |           |           |            |       |        |           |             |
|     Serialize_ModelWriter_Current |        8192 | 3,228.2 ns |  86.29 ns |  99.38 ns | 3,163.1 ns |  1.00 |      - |     728 B |        1.00 |
| Serialize_ModelWriter_Interlocked |        8192 |   756.2 ns |  15.89 ns |  18.30 ns |   750.0 ns |  0.23 | 0.0061 |     592 B |        0.81 |
|       Serialize_ModelWriter_Locks |        8192 |   799.2 ns |  17.50 ns |  20.15 ns |   786.3 ns |  0.25 | 0.0033 |     624 B |        0.86 |
|                                   |             |            |           |           |            |       |        |           |             |
|     Serialize_ModelWriter_Current |       16384 | 3,170.4 ns |  34.82 ns |  27.19 ns | 3,163.1 ns |  1.00 |      - |     728 B |        1.00 |
| Serialize_ModelWriter_Interlocked |       16384 |   736.5 ns |  10.49 ns |   8.76 ns |   734.5 ns |  0.23 | 0.0061 |     592 B |        0.81 |
|       Serialize_ModelWriter_Locks |       16384 |   784.2 ns |   2.35 ns |   1.96 ns |   784.4 ns |  0.25 | 0.0064 |     624 B |        0.86 |
|                                   |             |            |           |           |            |       |        |           |             |
|     Serialize_ModelWriter_Current |       32768 | 3,187.2 ns |  12.45 ns |  11.04 ns | 3,182.6 ns |  1.00 |      - |     728 B |        1.00 |
| Serialize_ModelWriter_Interlocked |       32768 |   746.6 ns |   5.99 ns |   5.60 ns |   745.4 ns |  0.23 | 0.0060 |     592 B |        0.81 |
|       Serialize_ModelWriter_Locks |       32768 |   774.7 ns |   3.32 ns |   2.77 ns |   774.3 ns |  0.24 | 0.0064 |     624 B |        0.86 |

ResourceProviderDataModel:
|                            Method | SegmentSize |     Mean |    Error |   StdDev | Ratio | RatioSD |   Gen0 |   Gen1 |   Gen2 | Allocated | Alloc Ratio |
|---------------------------------- |------------ |---------:|---------:|---------:|------:|--------:|-------:|-------:|-------:|----------:|------------:|
|     Serialize_ModelWriter_Current |        4096 | 619.1 us | 24.39 us | 28.09 us |  1.00 |    0.00 | 4.8077 | 4.8077 | 4.8077 | 198.89 KB |        1.00 |
| Serialize_ModelWriter_Interlocked |        4096 | 617.4 us | 26.48 us | 27.20 us |  0.99 |    0.07 | 4.8077 | 4.8077 | 4.8077 | 198.77 KB |        1.00 |
|       Serialize_ModelWriter_Locks |        4096 | 641.0 us | 16.15 us | 18.60 us |  1.04 |    0.05 | 5.2083 | 5.2083 | 5.2083 | 198.79 KB |        1.00 |
|                                   |             |          |          |          |       |         |        |        |        |           |             |
|     Serialize_ModelWriter_Current |        8192 | 626.2 us | 14.52 us | 16.14 us |  1.00 |    0.00 | 4.8077 | 4.8077 | 4.8077 | 181.22 KB |        1.00 |
| Serialize_ModelWriter_Interlocked |        8192 | 612.8 us | 15.94 us | 17.72 us |  0.98 |    0.04 | 4.6296 | 4.6296 | 4.6296 | 181.07 KB |        1.00 |
|       Serialize_ModelWriter_Locks |        8192 | 620.1 us | 21.28 us | 24.50 us |  0.99 |    0.05 | 5.0000 | 5.0000 | 5.0000 | 181.12 KB |        1.00 |
|                                   |             |          |          |          |       |         |        |        |        |           |             |
|     Serialize_ModelWriter_Current |       16384 | 624.8 us | 19.02 us | 21.91 us |  1.00 |    0.00 | 4.8077 | 4.8077 | 4.8077 | 176.45 KB |        1.00 |
| Serialize_ModelWriter_Interlocked |       16384 | 610.4 us | 23.69 us | 26.33 us |  0.98 |    0.06 | 4.8077 | 4.8077 | 4.8077 |  176.3 KB |        1.00 |
|       Serialize_ModelWriter_Locks |       16384 | 587.3 us |  7.09 us |  5.53 us |  0.94 |    0.03 | 4.4643 | 4.4643 | 4.4643 | 176.34 KB |        1.00 |
|                                   |             |          |          |          |       |         |        |        |        |           |             |
|     Serialize_ModelWriter_Current |       32768 | 625.4 us | 19.00 us | 21.88 us |  1.00 |    0.00 | 4.8077 | 4.8077 | 4.8077 | 175.24 KB |        1.00 |
| Serialize_ModelWriter_Interlocked |       32768 | 608.8 us | 20.40 us | 23.49 us |  0.97 |    0.06 | 5.0000 | 5.0000 | 5.0000 | 175.11 KB |        1.00 |
|       Serialize_ModelWriter_Locks |       32768 | 606.3 us | 19.09 us | 21.98 us |  0.97 |    0.04 | 5.0000 | 5.0000 | 5.0000 | 175.15 KB |        1.00 |

```